chore: move eip and bandwidth out of vpc directory
