+++ huaweicloud/provider.go
@@ -25,6 +25,7 @@ import (
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/dli"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/dms"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/dws"
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/eip"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/elb"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/eps"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/evs"
@@ -355,11 +356,11 @@ func Provider() *schema.Provider {
 			"huaweicloud_sfs_file_system":                      DataSourceSFSFileSystemV2(),
 			"huaweicloud_vbs_backup_policy":                    dataSourceVBSBackupPolicyV2(),
 			"huaweicloud_vbs_backup":                           dataSourceVBSBackupV2(),
+			"huaweicloud_vpc_bandwidth":                        eip.DataSourceBandWidth(),
+			"huaweicloud_vpc_eip":                              eip.DataSourceVpcEip(),
+			"huaweicloud_vpc_eips":                             eip.DataSourceVpcEips(),
 			"huaweicloud_vpc":                                  vpc.DataSourceVpcV1(),
 			"huaweicloud_vpcs":                                 vpc.DataSourceVpcs(),
-			"huaweicloud_vpc_bandwidth":                        vpc.DataSourceBandWidth(),
-			"huaweicloud_vpc_eip":                              vpc.DataSourceVpcEip(),
-			"huaweicloud_vpc_eips":                             vpc.DataSourceVpcEips(),
 			"huaweicloud_vpc_ids":                              vpc.DataSourceVpcIdsV1(),
 			"huaweicloud_vpc_peering_connection":               vpc.DataSourceVpcPeeringConnectionV2(),
 			"huaweicloud_vpc_route_table":                      vpc.DataSourceVPCRouteTable(),
@@ -580,12 +581,12 @@ func Provider() *schema.Provider {
 			"huaweicloud_swr_repository_sharing":          swr.ResourceSWRRepositorySharing(),
 			"huaweicloud_vbs_backup":                      resourceVBSBackupV2(),
 			"huaweicloud_vbs_backup_policy":               resourceVBSBackupPolicyV2(),
-			"huaweicloud_vpc":                             vpc.ResourceVirtualPrivateCloudV1(),
-			"huaweicloud_vpc_bandwidth":                   vpc.ResourceVpcBandWidthV2(),
-			"huaweicloud_vpc_eip":                         vpc.ResourceVpcEIPV1(),
+			"huaweicloud_vpc_bandwidth":                   eip.ResourceVpcBandWidthV2(),
+			"huaweicloud_vpc_eip":                         eip.ResourceVpcEIPV1(),
 			"huaweicloud_vpc_peering_connection":          vpc.ResourceVpcPeeringConnectionV2(),
 			"huaweicloud_vpc_peering_connection_accepter": vpc.ResourceVpcPeeringConnectionAccepterV2(),
 			"huaweicloud_vpc_route_table":                 vpc.ResourceVPCRouteTable(),
+			"huaweicloud_vpc":                             vpc.ResourceVirtualPrivateCloudV1(),
 			"huaweicloud_vpc_route":                       vpc.ResourceVPCRouteTableRoute(),
 			"huaweicloud_vpc_subnet":                      vpc.ResourceVpcSubnetV1(),
 			"huaweicloud_vpcep_approval":                  ResourceVPCEndpointApproval(),
@@ -649,12 +650,12 @@ func Provider() *schema.Provider {
 			"huaweicloud_sfs_file_system_v2":                 ResourceSFSFileSystemV2(),
 			"huaweicloud_iam_agency":                         iam.ResourceIAMAgencyV3(),
 			"huaweicloud_iam_agency_v3":                      iam.ResourceIAMAgencyV3(),
-			"huaweicloud_vpc_v1":                             vpc.ResourceVirtualPrivateCloudV1(),
-			"huaweicloud_vpc_bandwidth_v2":                   vpc.ResourceVpcBandWidthV2(),
-			"huaweicloud_vpc_eip_v1":                         vpc.ResourceVpcEIPV1(),
+			"huaweicloud_vpc_bandwidth_v2":                   eip.ResourceVpcBandWidthV2(),
+			"huaweicloud_vpc_eip_v1":                         eip.ResourceVpcEIPV1(),
 			"huaweicloud_vpc_peering_connection_v2":          vpc.ResourceVpcPeeringConnectionV2(),
 			"huaweicloud_vpc_peering_connection_accepter_v2": vpc.ResourceVpcPeeringConnectionAccepterV2(),
 			"huaweicloud_vpc_route_v2":                       vpc.ResourceVPCRouteV2(),
+			"huaweicloud_vpc_v1":                             vpc.ResourceVirtualPrivateCloudV1(),
 			"huaweicloud_vpc_subnet_v1":                      vpc.ResourceVpcSubnetV1(),
 			"huaweicloud_cce_cluster_v3":                     ResourceCCEClusterV3(),
 			"huaweicloud_cce_node_v3":                        ResourceCCENodeV3(),
