+++ huaweicloud/services/acceptance/eip/resource_huaweicloud_vpc_eip_test.go
@@ -1,4 +1,4 @@
-package vpc
+package eip
 
 import (
 	"fmt"
