+++ huaweicloud/services/eip/data_source_huaweicloud_vpc_eip.go
@@ -1,4 +1,4 @@
-package vpc
+package eip
 
 import (
 	"github.com/chnsz/golangsdk/openstack/networking/v1/eips"
