+++ huaweicloud/services/eip/data_source_huaweicloud_vpc_eips.go
@@ -1,4 +1,4 @@
-package vpc
+package eip
 
 import (
 	"context"
