+++ huaweicloud/services/eip/resource_huaweicloud_vpc_bandwidth.go
@@ -1,4 +1,4 @@
-package vpc
+package eip
 
 import (
 	"context"
