+++ huaweicloud/services/eip/resource_huaweicloud_vpc_eip.go
@@ -1,4 +1,4 @@
-package vpc
+package eip
 
 import (
 	"time"
